fix(review-rounds): drop redundant int() cast in percentile (ruff RUF046)
round() with a single argument already returns an int in Python 3, so the
int(...) wrapper was redundant. Behavior is identical; fixes the sole CI lint
error (RUF046) blocking #1242.

+++ scripts/review-rounds.py
@@ -302,7 +302,7 @@ def percentile(values, pct):
     if not values:
         return None
     s = sorted(values)
-    k = max(0, min(len(s) - 1, int(round((pct / 100.0) * (len(s) - 1)))))
+    k = max(0, min(len(s) - 1, round((pct / 100.0) * (len(s) - 1))))
     return s[k]
